billService refactor, part 2: no more billService
