+++ src/generate.ts
@@ -1,25 +1,45 @@
 import { aiService } from "./aiService/aiService";
-import { BillService } from "./billService/billService";
-import type { BillDoc, BillFull, DocSummary } from "./billService/types";
+import {
+  readBills,
+  readBillSummary,
+  writeBillFull,
+  writeBills,
+  writeBillSummary,
+} from "./bills/billStorage";
+import type {
+  Bill,
+  BillDoc,
+  BillFull,
+  BillSummary,
+  DocSummary,
+} from "./bills/types";
 import {
   getBillInfo,
   getLastActionDate,
   getLegislationIds,
-} from "./billService/wslHelpers";
+} from "./bills/wslHelpers";
+import { wslToBill, wslToBillFull } from "./bills/wslToBill";
 import { asyncBatch } from "./utils/asyncBatch";
 import { markdownToHtml } from "./utils/html";
 import { logger } from "./utils/logger";
 
-const billService = new BillService();
+const billMap = new Map<number, Bill>();
+let modified = false;
 
 export async function initialize() {
   logHeader("WA-Bill-TLDR Initialize");
-  return billService.initialize();
+  const bills = await readBills();
+  for (const bill of bills) {
+    billMap.set(bill.id, bill);
+  }
 }
 
 export async function finalize() {
   logHeader("Finalize");
-  await billService.saveBills();
+  if (modified) {
+    const bills = Array.from(billMap.values()).sort((a, b) => a.id - b.id);
+    await writeBills(bills);
+  }
   aiService.logAggregation();
   logHeader("WA-Bill-TLDR Complete");
 }
@@ -37,7 +57,8 @@ export async function findOutdatedIds(): Promise<number[]> {
 
   // Sequential execution with for...of (be kind to WSL's servers)
   for (const id of ids) {
-    const knownDate = billService.getLastUpdated(id);
+    const lastUpdated = billMap.get(id)?.lastUpdated;
+    const knownDate = lastUpdated ? new Date(lastUpdated) : undefined;
 
     if (!knownDate) {
       logger.debug(`Unknown bill ${id}`);
@@ -71,16 +92,26 @@ async function updateBill(id: number) {
   logger.info(`Updating bill ${id}`);
   const { legislation, sponsors, billDoc } = await getBillInfo(id);
 
-  return billService.upsertBill(legislation, sponsors, billDoc, [], []);
+  const bill = wslToBill(id, legislation, sponsors);
+
+  if (!bill) return;
+
+  const billFull = wslToBillFull(bill, billDoc, [], []);
+
+  billMap.set(id, bill);
+  modified = true;
+  await writeBillFull(id, billFull);
+
+  return billFull;
 }
 
 async function summarizeBill(bill: BillFull) {
   logger.info(`Summarizing bill ${bill.id}`);
-  const billSummary = await billService.getBillSummary(bill.id);
+  const billSummary = await getBillSummary(bill.id);
 
   await addSummaries(bill.id, bill.billDocuments, billSummary.documents);
 
-  await billService.saveBillSummary(billSummary);
+  await writeBillSummary(billSummary.id, billSummary);
 }
 
 async function addSummaries(
@@ -128,6 +159,18 @@ async function addSummary(
       };
     }
   }
+}
+
+async function getBillSummary(id: number): Promise<BillSummary> {
+  const summary = await readBillSummary(id);
+  return (
+    summary ?? {
+      id,
+      documents: {},
+      reports: {},
+      amendments: {},
+    }
+  );
 }
 
 function logHeader(msg: string) {
